Refactor LOOC and SubtleLOOC handling

+++ Content.Shared/Chat/ChatChannel.cs
@@ -4,7 +4,7 @@ namespace Content.Shared.Chat
     ///     Represents chat channels that the player can filter chat tabs by.
     /// </summary>
     [Flags]
-    public enum ChatChannel : ushort
+    public enum ChatChannel : uint
     {
         None = 0,
